chore: follow recommended practices - refactor into custom tasks
- Prefer custom task types over ad-hoc tasks. See https://docs.gradle.org/current/userguide/implementing_gradle_plugins.html#writing-and-using-custom-task-types
