+++ src/main/groovy/net/serenitybdd/plugins/gradle/ClearReportsTask.groovy
@@ -1,0 +1,15 @@
+package net.serenitybdd.plugins.gradle
+
+import org.apache.commons.io.FileUtils
+import org.gradle.api.tasks.TaskAction
+
+class ClearReportsTask extends SerenityAbstractTask {
+
+    @TaskAction
+    void clearReports() {
+        SerenityAbstractTask.updateProperties(project)
+        def reportDirectory = SerenityAbstractTask.prepareReportDirectory(project)
+        FileUtils.deleteDirectory(reportDirectory.toFile())
+    }
+
+}
